requirements_qa.txt
ast-comments==1.2.1
black==23.12.1
coverage[toml]==7.4.0
doc8==1.1.1
flake8-isort==6.1.1
flake8-pyproject==1.2.3
flake8==6.1.0
ghapi==1.0.4
httpx==0.26.0
mypy==1.8.0
pandas-stubs==2.1.4.231227
pur==7.3.1
pydantic==2.5.3
pydocstyle==6.3.0
pydoctest==0.1.22
pylint==2.17.7
pyspelling==2.9
pytest==7.4.4
regex==2023.12.25
resplendent==0.3.5
sphinx-rtd-theme==2.0.0
sphinx==7.2.6
sphinx_design==0.5.0
tqdm==4.66.1
typed-argument-parser==1.9.0
types-tqdm==4.66.0.5
